Move 'KeyEvent' into it's own class
